requirements.txt
Django==3.2.19
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simlejwt==5.2.2
drf-writable-nested==0.7.0
pyotp==2.8.0
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4 